+++ src/link_validator/file_system.rs
@@ -1,9 +1,10 @@
-use crate::Config;
 use crate::link_validator::LinkCheckResult;
-use async_std::path::Path;
-use std::path::MAIN_SEPARATOR;
+use crate::Config;
 use async_std::fs::canonicalize;
+use async_std::path::Path;
 use async_std::path::PathBuf;
+use std::path::MAIN_SEPARATOR;
+use walkdir::WalkDir;
 
 pub async fn check_filesystem(source: &str, target: &str, config: &Config) -> LinkCheckResult {
     let mut normalized_link = target
@@ -30,17 +31,57 @@ pub async fn check_filesystem(source: &str, target: &str, config: &Config) -> Li
     debug!("Check file system link target {:?}", target);
     let target = absolute_target_path(source, &fs_link_target).await;
     debug!("Absolute target path {:?}", target);
+    let error_result = LinkCheckResult::Failed("Target path not found.".to_string());
     if target.exists().await {
         LinkCheckResult::Ok
+    } else if !config.match_file_extension && target.extension().is_none() {
+        // Check if file exists ignoring the file extension
+        let target_file_name = match target.file_name() {
+            Some(s) => s,
+            None => return error_result,
+        };
+        let target_parent = match target.parent() {
+            Some(s) => s,
+            None => return error_result,
+        };
+        debug!("Check if file ignoring the extension exists.");
+        if target_parent.exists().await {
+            debug!(
+                "Parent {:?} exists. Search dir for file ignoring the extension.",
+                target_parent
+            );
+            for entry in WalkDir::new(target_parent)
+                .follow_links(false)
+                .max_depth(1)
+                .into_iter()
+                .filter_map(Result::ok)
+                .filter(|e| !e.file_type().is_dir())
+            {
+                let mut file_on_system = entry.into_path();
+                file_on_system.set_extension("");
+                match file_on_system.file_name() {
+                    Some(file_name) => {
+                        if target_file_name == file_name {
+                            info!("Found file {:?}", file_on_system);
+                            return LinkCheckResult::Ok;
+                        }
+                    }
+                    None => return error_result,
+                }
+            }
+            return error_result;
+        } else {
+            return error_result;
+        }
     } else {
-        LinkCheckResult::Failed("Target path not found.".to_string())
+        error_result
     }
 }
 
 async fn absolute_target_path(source: &str, target: &PathBuf) -> PathBuf {
     let abs_source = canonicalize(source).await.expect("Expected path to exist.");
     if target.is_relative() {
-        let parent = abs_source.parent().unwrap_or(Path::new("./"));
+        let parent = abs_source.parent().unwrap_or(Path::new("/"));
         parent.join(target)
     } else {
         target.to_owned()