+++ tests/end_to_end.rs
@@ -19,6 +19,7 @@ async fn end_to_end() {
         log_level: logger::LogLevel::Debug,
         markup_types: vec![MarkupType::Markdown],
         no_web_links: false,
+        match_file_extension: false,
         ignore_links: vec![],
         ignore_path: vec![
             fs::canonicalize("benches/benchmark/markdown/ignore_me.md").unwrap(),
@@ -45,6 +46,7 @@ async fn end_to_end_different_root() {
         log_level: logger::LogLevel::Debug,
         markup_types: vec![MarkupType::Markdown],
         no_web_links: false,
+        match_file_extension: false,
         ignore_links: vec![],
         ignore_path: vec![],
         root_dir: Some(test_files),